+++ index.html
@@ -57,6 +57,8 @@
     </div>
     
 
-    <script src="index.js"></script>
+    <script type="module" src="block_types.js"></script>
+    <script type="module" src="helpers.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 </html>
